WIP(restaurantScreen): detail page publishing
- wrap with CustomScrollView
- top info section
- menu list

Resolving: #7
