+++ lib/components/product_card.dart
@@ -1,0 +1,53 @@
+import 'package:delivery/styles/colors.dart';
+import 'package:flutter/material.dart';
+
+class ProductCard extends StatelessWidget {
+  const ProductCard({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return IntrinsicHeight(
+      child: Row(
+        children: [
+          ClipRRect(
+            borderRadius: BorderRadius.circular(8.0),
+            child: Image.asset(
+              'assets/images/food/ddeok_bok_gi.jpg',
+              width: 110,
+              height: 110,
+              fit: BoxFit.cover,
+            ),
+          ),
+          const SizedBox(width: 16.0),
+          Expanded(
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.stretch,
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                Text(
+                  '떡볶이',
+                  style: TextStyle(fontSize: 18.0, fontWeight: FontWeight.w500),
+                ),
+                Text(
+                  '전통 떡볶이의 정석!맛있습니다전통 떡볶이의 정석!맛있습니다전통 떡볶이의 정석!맛있습니다전통 떡볶이의 정석!맛있습니다',
+                  maxLines: 2,
+                  overflow: TextOverflow.ellipsis,
+                  style: TextStyle(color: C_BODY_TEXT, fontSize: 14.0),
+                ),
+                Text(
+                  '₩10000',
+                  textAlign: TextAlign.end,
+                  style: TextStyle(
+                    color: C_PRIMARY,
+                    fontSize: 12.0,
+                    fontWeight: FontWeight.w500,
+                  ),
+                ),
+              ],
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}
