+++ lib/screens/restaurant/_shared/restaurant_card.dart
@@ -1,0 +1,148 @@
+import 'package:delivery/models/restaurant/restaurant_model.dart';
+import 'package:delivery/styles/colors.dart';
+import 'package:flutter/material.dart';
+
+class RestaurantCard extends StatelessWidget {
+  final Widget imageWidget;
+  final String name;
+  final List<String> tags;
+  final int ratingsCount;
+  final int deliveryTime;
+  final int deliveryFee;
+  final double ratings;
+  final bool isDetail;
+  final String? detail;
+
+  const RestaurantCard({
+    super.key,
+    required this.imageWidget,
+    required this.name,
+    required this.tags,
+    required this.ratingsCount,
+    required this.deliveryFee,
+    required this.deliveryTime,
+    required this.ratings,
+    this.isDetail = false,
+    this.detail,
+  });
+
+  factory RestaurantCard.fromModel({
+    required RestaurantModel model,
+    bool isDetail = false,
+  }) {
+    return RestaurantCard(
+      imageWidget: Image.network(
+        model.thumbUrl,
+        fit: BoxFit.cover,
+      ),
+      name: model.name,
+      tags: model.tags,
+      ratingsCount: model.ratingsCount,
+      ratings: model.ratings,
+      deliveryTime: model.deliveryTime,
+      deliveryFee: model.deliveryFee,
+      isDetail: isDetail,
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: [
+        if (isDetail) imageWidget,
+        if (!isDetail)
+          ClipRRect(
+            borderRadius: BorderRadius.circular(12.0),
+            child: imageWidget,
+          ),
+        const SizedBox(height: 16.0),
+        Padding(
+          padding: EdgeInsets.symmetric(horizontal: isDetail ? 16.0 : 0),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.stretch,
+            children: [
+              Text(
+                name,
+                style: const TextStyle(
+                  fontSize: 20.0,
+                  fontWeight: FontWeight.w500,
+                ),
+              ),
+              const SizedBox(height: 8.0),
+              Text(
+                tags.join(' · '),
+                style: const TextStyle(
+                  color: C_BODY_TEXT,
+                  fontSize: 14.0,
+                ),
+              ),
+              const SizedBox(height: 8.0),
+              Row(
+                children: [
+                  _IconText(
+                    icon: Icons.star,
+                    label: ratings.toString(),
+                  ),
+                  renderDot(),
+                  _IconText(
+                    icon: Icons.receipt,
+                    label: ratingsCount.toString(),
+                  ),
+                  renderDot(),
+                  _IconText(
+                    icon: Icons.timelapse_outlined,
+                    label: '$deliveryTime분',
+                  ),
+                  renderDot(),
+                  _IconText(
+                    icon: Icons.monetization_on,
+                    label: deliveryFee == 0 ? '무료' : deliveryFee.toString(),
+                  ),
+                ],
+              ),
+              if (detail != null && isDetail)
+                Padding(
+                  padding: const EdgeInsets.symmetric(vertical: 16.0),
+                  child: Text(detail!),
+                )
+            ],
+          ),
+        )
+      ],
+    );
+  }
+
+  Widget renderDot() {
+    return const Padding(
+      padding: EdgeInsets.symmetric(horizontal: 4.0),
+      child: Text(
+        '·',
+        style: TextStyle(fontSize: 12.0, fontWeight: FontWeight.w500),
+      ),
+    );
+  }
+}
+
+class _IconText extends StatelessWidget {
+  final IconData icon;
+  final String label;
+
+  const _IconText({super.key, required this.label, required this.icon});
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        Icon(icon, color: C_PRIMARY, size: 14.0),
+        const SizedBox(width: 8.0),
+        Text(
+          label,
+          style: const TextStyle(
+            fontSize: 12.0,
+            fontWeight: FontWeight.w500,
+          ),
+        ),
+      ],
+    );
+  }
+}
