--- lib/screens/restaurant/detail/restarurant_detail_screen.dart
@@ -1,16 +1,0 @@
-import 'package:delivery/components/layout/default_layout.dart';
-import 'package:flutter/material.dart';
-
-class RestaurantDetailScreen extends StatelessWidget {
-  const RestaurantDetailScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return DefaultLayout(
-      title: '불타는 떡볶이',
-      child: Column(
-        children: [],
-      ),
-    );
-  }
-}
